Implemented Cart, bugfixes

+++ controllers/Cart.js
@@ -6,25 +6,23 @@ var session = require('../utils/SessionManager');
 
 
 module.exports.cartGET = function cartGET (req, res, next) {
+    var limit = req.swagger.params['limit'].value;
     var offset = req.swagger.params['offset'].value;
 
-    //TODO se la sessione è valida, recupero userId e chiamo il service
-    //TODO IMPLEMENTARE
-    if(false){
-        var userId = 1;
-        Cart.cartGET(userId, offset)
-            .then(function (response) {
-                utils.writeJson(res, response);
-                //TODO salvare il carrello nella sessione
-            })
-            .catch(function (response) {
-                utils.writeJson(res, response);
-            });
-    } else {
-        let response = {message: "Not authorized!"};
-        utils.writeJson(res, response, 403);
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
     }
 
+    let userId = session.getUserId();
+
+    Cart.cartGET(userId, limit, offset)
+        .then(function (response) {
+            utils.writeJson(res, response);
+        })
+        .catch(function (response) {
+            utils.writeJson(res, response, response.errorCode || 500);
+        });
 
 };
 
@@ -32,13 +30,19 @@ module.exports.cartRemoveDELETE = function cartRemoveDELETE (req, res, next) {
     var bookId = req.swagger.params['bookId'].value;
     var quantity = req.swagger.params['quantity'].value;
 
-    //TODO IMPLEMENTARE
-    Cart.cartRemoveDELETE(bookId,quantity)
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+    Cart.cartRemoveDELETE(userId, bookId,quantity)
         .then(function (response) {
             utils.writeJson(res, response);
         })
         .catch(function (response) {
-            utils.writeJson(res, response);
+            utils.writeJson(res, response, response.errorCode || 500);
         });
 };
 
@@ -46,12 +50,19 @@ module.exports.cartUpdatePUT = function cartUpdatePUT (req, res, next) {
     var bookId = req.swagger.params['bookId'].value;
     var quantity = req.swagger.params['quantity'].value;
 
-    //TODO IMPLEMENTARE
-    Cart.cartUpdatePUT(bookId,quantity)
+    if(!session.userIdExists()){
+        utils.writeJson(res, {message: "You must login to perform this operation!"}, 403);
+        return;
+    }
+
+    let userId = session.getUserId();
+
+    Cart.cartUpdatePUT(userId, bookId,quantity)
         .then(function (response) {
             utils.writeJson(res, response);
         })
         .catch(function (response) {
-            utils.writeJson(res, response);
+            console.log(response);
+            utils.writeJson(res, response, response.errorCode || 500);
         });
 };
